refactor: 알림 개수 타입을 Integer에서 Long으로 변경하여 불필요한 형변환을 제거했습니다

+++ src/main/java/com/example/cherrydan/activity/dto/UnreadAlertCountResponseDTO.java
@@ -11,11 +11,11 @@ import lombok.Getter;
 public class UnreadAlertCountResponseDTO {
 
     @Schema(description = "총 미읽은 알림 개수", example = "15", nullable = false, requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer totalCount;
+    private Long totalCount;
 
     @Schema(description = "활동 알림 미읽은 개수", example = "8", nullable = false, requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer activityAlertCount;
+    private Long activityAlertCount;
 
     @Schema(description = "키워드 알림 미읽은 개수", example = "7", nullable = false, requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer keywordAlertCount;
+    private Long keywordAlertCount;
 }
